fix(gpu): 32-bit flat indexing made the col-major PCA kernels fail past 2.1e9 elements (#423)
* fix(gpu): 32-bit flat indexing made the col-major PCA kernels fail past 2.1e9 elements

`scale_columns_kernel` and `mean_correct_colmajor_strided_kernel` computed
`int total = m * k` and indexed with a 32-bit `idx`, while the host launched a
grid sized in `usize`. `n_obs × k` crosses 2^31 at census scale — 36M cells ×
k=60 = 2.16e9 — and `randomized_pca_core`'s VRAM pre-flight admits that shape
on an 80 GB H100, so this was reachable, not theoretical.

`int total = m * k` is signed overflow, i.e. UB, so the manifestation is the
compiler's choice. Measured against the pre-fix kernels at 2^31 + 256 elements
(nvcc 12.6, -O3, compute_70, H100): CUDA_ERROR_ILLEGAL_ADDRESS — threads whose
`idx` wrapped negative survive the bounds check and write below the buffer,
which poisons the CUDA context for the process. Past 2^32 elements the host's
`(total as u32)` block count is the silent variant instead: a grid too small to
reach the tail, no error at all. Either way the work does not happen, on the
two paths that matter — `zero_center=True` yielding an uncentered PCA
(gpu_pca_resident.rs passes shard_rows = n_obs), and the final embedding
leaving unscaled by its singular values.

Widen the three col-major kernels to `long long` parameters with 64-bit totals
and flat indices, matching spmm_mean_correct.cu, the row-major sibling that
always had this right. `outer_sub_kernel` gets the same treatment; n_vars × k
is ~3e6 today, so that one is defence-in-depth.

Delete `select_top_eigvecs_desc_kernel` and `reverse_tail_vec_kernel`: dead
since the GPU covariance-PCA path was removed, and the first carried the
identical overflow. A dead broken kernel is worth less than the diff to fix it.

Host side, add `scx_gpu::flat_launch_1d(total: u64, threads: u32)` — the
extraction of the one already-correct copy of this arithmetic, in
`mean_correct_gpu` — and route five truncating `(total as u32).div_ceil(...)`
sites through it. It counts in u64 and returns `ShapeMismatch` when the block
count exceeds the driver's `grid_dim.x` cap rather than casting a wrapped value.

Also promote the two warp-per-row ids the review names as latent siblings
(sparse_dense.cu, normalize_log1p.cu::normalize_warp_row) and
`row_scale_kernel`'s global `row_offset + row`. Those launch 32 threads per row,
so their flat index wraps at 134.2M rows; per-shard row counts are far below
that, so this is defence-in-depth and the comments say so.

Addresses review §8.4.

* fix(gpu): address round-1 review — VRAM gate default, row_offset ABI, zero-threads guard

Cursor Agent - Grok 4.6 High Fast:
- `SCX_REQUIRE_LARGE_VRAM` was opt-in, so the one test that can observe the
  2^31 kernel overflow could skip on a default harness run and the suite would
  still go green. `_run_scx_gpu_tests.sh` now defaults it on for the same reason
  it already forces SCX_REQUIRE_GPU=1 — the job holds a whole H100 via
  --gres=gpu:1, so the ~10 GiB headroom is a property of the allocation.
  Override with SCX_REQUIRE_LARGE_VRAM=0 on a smaller or shared GPU.
- `gpu_hvg.rs` was named as an uncovered site: its six launchers counted in u64
  but cast the block count to u32 with no grid_dim.x cap check. All six now go
  through `flat_launch_1d`, which makes the crate's 1-D launch geometry uniform
  (15 sites).

gemini-code-assist:
- `row_scale_kernel`'s comment claimed the kernel was "correct on its own terms
  rather than by the caller's grace", but `row_offset` was still an `int`
  PARAMETER — an out-of-range value truncated at the ABI boundary before the
  promoted addition saw it, so the host's i32::MAX guard was still what made it
  safe. Widen the parameter to `long long`, pass i64, and drop the now-pointless
  host cap; the `factors.len()` bounds guard above it is the real check.
- `flat_launch_1d` validated `threads > 0` with a `debug_assert!`, which is
  compiled out in release, leaving `div_ceil` to panic on a bare divide-by-zero.
  It returns a Result and is re-exported from the crate root, so make it a real
  check. New test watched failing against the assert form first.

codex - gpt-5.6-sol:
- The new boundary test used `CudaStream::memcpy_dtov`, deprecated in cudarc in
  favour of `clone_dtoh`. CI's clippy does not lint test targets so it passed,
  but `cargo clippy -p scx-gpu --lib --tests` flagged it. Switched.

PR body corrected in place with visible correction blocks for the two claims
this round proved wrong.

* docs(gpu): address round-2 review — all four findings are doc/comment accuracy

No logic changed. Every finding this round was a claim that did not match the
code it described.

codex - gpt-5.6-sol:
- The kernel comments and PR body concluded "either way the PCA runs
  uncentered", flattening two regimes that fail differently. Only the >2^32
  host under-launch completes and returns a wrong result; at the measured 2^31
  shape the call FAILS (illegal access at synchronize, CUDA context poisoned)
  and there is no PCA result to be uncentered. Split them explicitly.
- `flat_launch_1d`'s doc said the cap check makes "a matrix past 2^31 elements
  error". It does not, and the test right below asserts the opposite: the
  rejection is on the block count, needing total > 256 * (2^31 - 1) ~= 5.5e11
  elements. The u64 count is what fixes the >2^32 truncation; the cap is a
  separate, much higher guard.

Cursor Agent - Grok 4.6 High Fast:
- `docs/conventions.md` named the wrong product for the 36M crossing.
  36M x 50 PCs = 1.80e9, still under 2^31; the 2.16e9 figure is n_obs x k with
  k = n_components + n_oversamples. n_obs x n_comps crosses around 43M cells.
  The kernel comments had the split right; the convention meant to stop the
  next kernel repeating the bug did not.
- The skip-summary footer hardcoded SCX_REQUIRE_LARGE_VRAM=1, so under the
  documented =0 override a free-VRAM skip would have been labelled an
  nvcomp/cuVS skip. It now branches on the actual value.
- `docs/testing.md` and `test_gate.rs`'s module doc still described the harness
  as setting only SCX_REQUIRE_GPU=1, drift introduced by the previous commit
  making the VRAM default load-bearing. Both updated.

PR body carries visible correction blocks for all three claims of its own that
this round proved wrong.

+++ scx-gpu/kernels/normalize_log1p.cu
@@ -93,11 +93,17 @@ extern "C" __global__ void log1p_kernel(
 // factor is supplied. `factors` is a global per-row vector indexed in the
 // source's iteration row order; `row_offset` is this shard's first global
 // row, so one device factor vector serves every shard in a streamed source.
+// `row_offset` is `long long` (review §8.4): `factors` is indexed in the
+// *global* obs row space, so `row_offset + row` is an n_obs-scale quantity even
+// though each shard's `n_rows` is small. Promoting only the addition would have
+// left the value truncating at the parameter boundary instead — the host would
+// still have had to cap `row_offset` at i32::MAX for the kernel to be safe,
+// which is a cap on the obs axis with no reason to exist.
 extern "C" __global__ void row_scale_kernel(
     const long long* __restrict__ indptr,
     float* __restrict__ data,
     int n_rows,
-    int row_offset,
+    long long row_offset,
     const float* __restrict__ factors
 ) {
     int row = blockIdx.x * blockDim.x + threadIdx.x;
@@ -189,8 +195,13 @@ __device__ __forceinline__ void normalize_warp_row(
     bool do_log1p
 ) {
     int lane = threadIdx.x % warpSize;
-    int warp_global = (blockIdx.x * blockDim.x + threadIdx.x) / warpSize;
-    if (warp_global >= n_rows) return;
+    // 64-bit warp id (review §8.4): one warp per row means the flat thread index
+    // is 32 × n_rows, which wraps `unsigned` past 134.2M rows; the negative id
+    // then passes the `>= n_rows` guard and `indptr[warp_global]` reads out of
+    // bounds. Per-shard row counts are far below that, so this is
+    // defence-in-depth — the nonzero addressing below is already 64-bit.
+    long long warp_global = ((long long)blockIdx.x * blockDim.x + threadIdx.x) / warpSize;
+    if (warp_global >= (long long)n_rows) return;
 
     long long start = indptr[warp_global];
     long long end = indptr[warp_global + 1];
